requirements.txt
python-telegram-bot==20.6
pymongo==4.6.1
requests==2.31.0
python-dotenv==1.0.0
apscheduler==3.10.1
imap-tools==1.4.0
python-dateutil==2.8.2